Major change to CriteriaParameter handling
Revamped Criteria class so that all criteria datatypes match
those coming right out of CriteriaParameter parsing.
  - starttime/endtime are now datetime objects
  - duration is a timedelta object

CriteriaParameters are attached either to a table or a report, but
should be attached to any table that can be run on its own (via
run_table).

Fixed run_table to take -C '<keyword>:<value>' for all criteria.

Model for criteria is now that each criteria should be parseable
as a string in unit, thus cannot use MultiField style fields
because that's a pain trying to parse from the command line.

Parsing of command line as well as parameters for a UI report
are now all CriteriaParameter driven.