More criteria work, TableField rename, report sections
Renamed CriteriaParameter to TableField to avoid the confusion between
running criteria as a set of values and a single field or set of
fields that define what a user inputs.

Implemented report Sections -- all tables in the same section
get the same criteria.  Mutliple sections can be used to get
different fields in the user input for the same eventual criteria
keywords.