More criteria revamp; eliminated default duration; section variables; file upload
Fixed file upload with the new criteria model -- this uses a new field
whose underlying datatype is a string instead of a file and accepts
an uploaded file.  This will work well later when going to a file section
model.

Eliminated the "default" duration because it was confusing.  Instead,
this can be implemented by making different sections with duration
local to that section.  This will show up in the criteria as
multiple duration selections (unless marked as hidden).

Support for ordering fields at the report level for presentation
purposese.

Tests now call both the report via the test client as well as
running individual tables via the run_table management command.